Vision/3.baseline/vision_baseline_VERJava/requirements.txt
GitPython==3.1.43
GitPython==3.1.43
tqdm==4.66.4
tree_sitter==0.22.3
tree_sitter_java==0.21.0
